--- app/src/main/java/com/example/imdbtestapp/TmdbApplication.kt
@@ -1,20 +1,0 @@
-package com.example.imdbtestapp
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-import timber.log.Timber
-
-@HiltAndroidApp
-class TmdbApplication : Application() {
-
-    override fun onCreate() {
-        super.onCreate()
-        initTimber()
-    }
-
-    private fun initTimber() {
-        if (BuildConfig.DEBUG) {
-            Timber.plant(Timber.DebugTree())
-        }
-    }
-}
