--- app/src/main/java/com/example/imdbtestapp/data/RemoteKeysDao.kt
@@ -1,20 +1,0 @@
-package com.example.imdbtestapp.data
-
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.OnConflictStrategy
-import androidx.room.Query
-import com.example.imdbtestapp.data.models.DbRemoteKeys
-
-@Dao
-interface RemoteKeysDao {
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertAll(remoteKey: List<DbRemoteKeys>)
-
-    @Query("SELECT * FROM remote_keys WHERE movieId = :movieId")
-    suspend fun remoteKeyMovieId(movieId: Long): DbRemoteKeys?
-
-    @Query("DELETE FROM remote_keys")
-    suspend fun clearRemoteKeys()
-}
