--- app/src/main/java/com/example/imdbtestapp/data/models/DbRemoteKeys.kt
@@ -1,12 +1,0 @@
-package com.example.imdbtestapp.data.models
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity(tableName = "remote_keys")
-data class DbRemoteKeys(
-    @PrimaryKey
-    val movieId: Long,
-    val prevKey: Int?,
-    val nextKey: Int?
-)
