--- app/src/main/java/com/example/imdbtestapp/mappers/MovieMapper.kt
@@ -1,26 +1,0 @@
-package com.example.imdbtestapp.mappers
-
-import com.example.imdbtestapp.data.models.DbMovie
-import com.example.imdbtestapp.models.Movie
-
-fun DbMovie.toApiMovie() =
-    Movie(
-        id = this.id,
-        originalLanguage = this.originalLanguage,
-        originalTitle = this.title,
-        overview = this.overview,
-        posterPath = this.posterPath,
-        releaseDate = this.releaseDate,
-        title = this.title
-    )
-
-fun Movie.toDbMovie() =
-    DbMovie(
-        id = this.id,
-        originalLanguage = this.originalLanguage,
-        originalTitle = this.title,
-        overview = this.overview,
-        posterPath = this.posterPath,
-        releaseDate = this.releaseDate,
-        title = this.title
-    )
