--- app/src/main/java/com/example/imdbtestapp/models/Movie.kt
@@ -1,19 +1,0 @@
-package com.example.imdbtestapp.models
-
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class Movie(
-    val id: Long,
-    @Json(name = "original_language")
-    val originalLanguage: String?,
-    @Json(name = "original_title")
-    val originalTitle: String?,
-    val overview: String?,
-    @Json(name = "poster_path")
-    val posterPath: String?,
-    @Json(name = "release_date")
-    val releaseDate: String?,
-    val title: String?
-)
