--- app/src/main/java/com/example/imdbtestapp/utils/HttpClient.kt
@@ -1,15 +1,0 @@
-package com.example.imdbtestapp.utils
-
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import java.util.concurrent.TimeUnit
-
-object HttpClient {
-    fun create(httpLoggingInterceptor: HttpLoggingInterceptor): OkHttpClient {
-        return OkHttpClient.Builder()
-            .addInterceptor(httpLoggingInterceptor)
-            .connectTimeout(API_CONNECT_TIMEOUT, TimeUnit.SECONDS)
-            .readTimeout(API_READ_TIMEOUT, TimeUnit.SECONDS)
-            .build()
-    }
-}
