--- app/src/main/java/com/example/imdbtestapp/utils/LoggingInterceptor.kt
@@ -1,10 +1,0 @@
-package com.example.imdbtestapp.utils
-
-import okhttp3.logging.HttpLoggingInterceptor
-
-object LoggingInterceptor {
-
-    fun create(): HttpLoggingInterceptor = HttpLoggingInterceptor().apply {
-        level = HttpLoggingInterceptor.Level.BODY
-    }
-}
