--- app/src/main/java/com/example/imdbtestapp/utils/MoshiBuilder.kt
@@ -1,10 +1,0 @@
-package com.example.imdbtestapp.utils
-
-import com.squareup.moshi.Moshi
-import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
-
-object MoshiBuilder {
-    fun create(): Moshi = Moshi.Builder()
-        .addLast(KotlinJsonAdapterFactory())
-        .build()
-}
