+++ app/src/main/java/com/example/tmdbtestapp/TmdbApplication.kt
@@ -1,0 +1,20 @@
+package com.example.tmdbtestapp
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+import timber.log.Timber
+
+@HiltAndroidApp
+class TmdbApplication : Application() {
+
+    override fun onCreate() {
+        super.onCreate()
+        initTimber()
+    }
+
+    private fun initTimber() {
+        if (BuildConfig.DEBUG) {
+            Timber.plant(Timber.DebugTree())
+        }
+    }
+}
