+++ app/src/main/java/com/example/tmdbtestapp/data/models/DbRemoteKeys.kt
@@ -1,0 +1,12 @@
+package com.example.tmdbtestapp.data.models
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity(tableName = "remote_keys")
+data class DbRemoteKeys(
+    @PrimaryKey
+    val movieId: Long,
+    val prevKey: Int?,
+    val nextKey: Int?
+)
