+++ app/src/main/java/com/example/tmdbtestapp/repo/MovieRepository.kt
@@ -1,0 +1,10 @@
+package com.example.tmdbtestapp.repo
+
+import androidx.paging.PagingData
+import com.example.tmdbtestapp.data.models.DbMovie
+import kotlinx.coroutines.flow.Flow
+
+interface MovieRepository {
+
+    fun getMovies(): Flow<PagingData<DbMovie>>
+}
