+++ app/src/main/java/com/example/tmdbtestapp/ui/MovieImage.kt
@@ -1,0 +1,28 @@
+package com.example.tmdbtestapp.ui
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.painterResource
+import coil.compose.AsyncImage
+import coil.request.ImageRequest
+import com.example.tmdbtestapp.R
+
+@Composable
+fun MovieImage(
+    imageUrl: String,
+    modifier: Modifier = Modifier
+) {
+    AsyncImage(
+        model = ImageRequest.Builder(LocalContext.current)
+            .data(imageUrl)
+            .crossfade(true)
+            .build(),
+        placeholder = painterResource(R.drawable.ic_photo), alpha = 0.45f,
+        contentDescription = null,
+        contentScale = ContentScale.Crop,
+        modifier = modifier
+    )
+
+}
