+++ app/src/main/java/com/example/tmdbtestapp/utils/Constants.kt
@@ -1,0 +1,8 @@
+package com.example.tmdbtestapp.utils
+
+const val TMDB_BASE_URL = "https://api.themoviedb.org/3/"
+const val TMDB_IMAGE_BASE_URL = "https://image.tmdb.org/t/p/w500"
+const val TMDB_STARTING_PAGE_INDEX = 1
+const val NETWORK_PAGE_SIZE = 30
+const val API_CONNECT_TIMEOUT = 300L
+const val API_READ_TIMEOUT = 300L
