+++ app/src/main/java/com/example/tmdbtestapp/utils/LoggingInterceptor.kt
@@ -1,0 +1,10 @@
+package com.example.tmdbtestapp.utils
+
+import okhttp3.logging.HttpLoggingInterceptor
+
+object LoggingInterceptor {
+
+    fun create(): HttpLoggingInterceptor = HttpLoggingInterceptor().apply {
+        level = HttpLoggingInterceptor.Level.BODY
+    }
+}
