+++ app/src/main/java/com/example/tmdbtestapp/utils/MoshiBuilder.kt
@@ -1,0 +1,10 @@
+package com.example.tmdbtestapp.utils
+
+import com.squareup.moshi.Moshi
+import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
+
+object MoshiBuilder {
+    fun create(): Moshi = Moshi.Builder()
+        .addLast(KotlinJsonAdapterFactory())
+        .build()
+}
