+++ settings.gradle
@@ -12,5 +12,5 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "ImdbTestApp"
+rootProject.name = "TmdbTestApp"
 include ':app'
